update form frmChiTietToaThuoc

+++ PhongKham/Forms_DE/frmToaThuoc.cs
@@ -6,9 +6,11 @@ using System.ComponentModel;
 using System.Data;
 using System.Drawing;
 using System.Linq;
+using System.Runtime.Remoting.Contexts;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using System.Windows.Input;
 
 namespace PhongKham.Forms_DE
 {
@@ -21,7 +23,15 @@ namespace PhongKham.Forms_DE
 
         private void frmToaThuoc_Load(object sender, EventArgs e)
         {
+            // TODO: This line of code loads data into the 'phongKhamDataSet.tToaThuoc' table. You can move, or remove it, as needed.
             toolStripDelete.Enabled = false;
+            btnEditChiTietToaThuoc.Enabled = false;
+            dtNgayKham.EditValue = DateTime.Now;
+            txtIdToaThuoc.Properties.ReadOnly = true;
+            lookUpEdit1.EditValue = "";
+            //lookUpEdit1.Properties.SearchMode = DevExpress.XtraEditors.Controls.SearchMode.AutoSuggest;
+
+
             loadData();
             loadCombobox();
         }
@@ -29,10 +39,21 @@ namespace PhongKham.Forms_DE
         void loadCombobox()
         {
             var tbenhNhans = new PhongKhamEntities().tBenhNhans.ToList();
-            cbBenhNhan.DisplayMember = "TenBenhNhan";
-            cbBenhNhan.ValueMember = "MaBenhNhan";
-            cbBenhNhan.DataSource = tbenhNhans;
-            cbBenhNhan.SelectedIndex = -1;
+
+            DataTable table = new DataTable();
+            table.Columns.Add("Mã bệnh nhân", typeof(string));
+            table.Columns.Add("Tên bệnh nhân", typeof(string));
+            table.Columns.Add("Tuổi", typeof(string));
+
+            foreach (var r in tbenhNhans)
+            {
+                table.Rows.Add(r.MaBenhNhan, r.TenBenhNhan, r.Tuoi);
+            }
+
+            lookUpEdit1.Properties.DisplayMember = "Tên bệnh nhân";
+            lookUpEdit1.Properties.ValueMember = "Mã bệnh nhân";
+            
+            lookUpEdit1.Properties.DataSource = table;
         }
 
         void loadData()
@@ -42,23 +63,27 @@ namespace PhongKham.Forms_DE
                 var tToaThuocs = from t in context.tToaThuocs
                                  select t;
 
+
                 DataTable table = new DataTable();
-                table.Columns.Add("STT", typeof(int));
+                table.Columns.Add("STT toa thuốc", typeof(int));
                 table.Columns.Add("Bệnh nhân", typeof(string));
                 table.Columns.Add("Bệnh được chẩn đoán", typeof(string));
                 table.Columns.Add("Ngày khám", typeof(string));
+                table.Columns.Add("Mã bệnh nhân", typeof(string));
 
                 foreach (var t in tToaThuocs)
                 {
                     string ngaykham = ((DateTime)t.NgayKham).ToString("MMMM dd yyyy");
-                    table.Rows.Add(t.STT, t.tBenhNhan.TenBenhNhan, t.BenhDuocChanDoan, ngaykham);
+                    table.Rows.Add(t.STT, t.tBenhNhan.TenBenhNhan, t.BenhDuocChanDoan, ngaykham, t.tBenhNhan.MaBenhNhan);
                 }
 
                 dgvToaThuoc.DataSource = table;
-                gridControlBenhNhan.DataSource = table;
+
             }
         }
 
+
+        //####################  CELL CLICK #################################################
         private void dgvToaThuoc_CellClick(object sender, DataGridViewCellEventArgs e)
         {
             if (e.RowIndex < 0 || e.RowIndex == dgvToaThuoc.Rows.Count - 1)
@@ -67,18 +92,44 @@ namespace PhongKham.Forms_DE
                 return;
             }
 
+            btnEditChiTietToaThuoc.Enabled = true;
+
             //DataGridViewRow row = dgvBenhNhan.Rows[dgvBenhNhan.SelectedCells[0].RowIndex];
             DataGridViewRow row = dgvToaThuoc.Rows[e.RowIndex];
             string gia = row.Cells[3].Value.ToString();
+            string idToaThuoc = row.Cells[0].Value.ToString();
 
-            txtIdToaThuoc.Text = row.Cells[0].Value.ToString();
-            cbBenhNhan.Text = row.Cells[1].Value.ToString();
+            txtIdToaThuoc.Text = idToaThuoc;
+            lookUpEdit1.EditValue = row.Cells[4].Value.ToString();
             txtBenhDuocChanDoan.Text = row.Cells[2].Value.ToString();
-            dtNgayKham.Value = DateTime.Parse(row.Cells[3].Value.ToString());
-
+            dtNgayKham.EditValue = DateTime.Parse(row.Cells[3].Value.ToString());
 
             //enable controls
             toolStripDelete.Enabled = true;
+
+            //load tChiTietToaThuocs
+            loadChiTietToaThuoc();
+
+        }
+
+        private void loadChiTietToaThuoc()
+        {
+            var db = new PhongKhamEntities();
+            var tChiTietToaThuocs = from record in db.tChiTietToaThuocs
+                                    where record.STT.ToString() == txtIdToaThuoc.Text.ToString()
+                                    select record;
+
+            DataTable table = new DataTable();
+            table.Columns.Add("Thuốc", typeof(string));
+            table.Columns.Add("Số lượng", typeof(string));
+            table.Columns.Add("Liều dùng", typeof(string));
+
+            foreach (var r in tChiTietToaThuocs)
+            {
+                table.Rows.Add(r.tThuoc.TenThuoc, r.SoLuong, r.LieuDung);
+            }
+
+            dgvChiTietToaThuoc.DataSource = table;
         }
 
         // ################################## SAVE #######################################
@@ -87,7 +138,7 @@ namespace PhongKham.Forms_DE
             if (txtIdToaThuoc.Text.Length == 0)
                 themToatThuoc();
             else
-                CapNhatToaThuoc();
+                capNhatToaThuoc();
         }
 
         // ################################## CANCEL #######################################
@@ -95,43 +146,56 @@ namespace PhongKham.Forms_DE
         {
             txtIdToaThuoc.Text = "";
             txtBenhDuocChanDoan.Text = "";
-            dtNgayKham.Value = DateTime.Now;
-            cbBenhNhan.SelectedIndex = -1;
+            dtNgayKham.EditValue = DateTime.Now;
             toolStripDelete.Enabled = false;
+            dgvChiTietToaThuoc.DataSource = null;
+            btnEditChiTietToaThuoc.Enabled = false;
+            lookUpEdit1.EditValue = "";
         }
 
         // ################################## DELETE #######################################
         private void toolStripDelete_Click(object sender, EventArgs e)
         {
-            using (var context = new PhongKhamEntities())
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn xóa?", "Xóa tóa thuốc", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
             {
-                var stub = new tToaThuoc { STT = Convert.ToInt32(txtIdToaThuoc.Text) };
-                context.tToaThuocs.Attach(stub);
-                context.tToaThuocs.Remove(stub);
-                context.SaveChanges();
-            }
+                using (var context = new PhongKhamEntities())
+                {
+                    var stub = new tToaThuoc { STT = Convert.ToInt32(txtIdToaThuoc.Text) };
+                    context.tToaThuocs.Attach(stub);
+                    context.tToaThuocs.Remove(stub);
+                    context.SaveChanges();
+                }
 
-            MessageBox.Show("Đã xóa toa thuốc" + txtIdToaThuoc.Text);
-            toolStripDelete.Enabled = false;
-            toolStripCancel.PerformClick();
-            loadData();
+                MessageBox.Show("Đã xóa toa thuốc" + txtIdToaThuoc.Text);
+                toolStripDelete.Enabled = false;
+                toolStripCancel.PerformClick();
+                loadData();
+            }
+            else if (dialogResult == DialogResult.No) {}
         }
 
         // ################################## ADD #######################################
         void themToatThuoc()
         {
-            if (txtBenhDuocChanDoan.Text.Length == 0 ||
-                cbBenhNhan.SelectedIndex == -1)
+            if (txtBenhDuocChanDoan.Text.Length == 0)
             {
                 MessageBox.Show("Hãy nhập đủ thông tin.");
+                return;
+            }
+
+            if (lookUpEdit1.EditValue == "")
+            {
+                MessageBox.Show("Hãy chọn bệnh nhân.");
                 return;
             }
 
             tToaThuoc toaThuoc = new tToaThuoc
             {
                 BenhDuocChanDoan = txtBenhDuocChanDoan.Text,
-                BenhNhan = cbBenhNhan.SelectedValue.ToString(),
-                NgayKham = dtNgayKham.Value,
+                //BenhNhan = cbBenhNhan.SelectedValue.ToString(),
+                BenhNhan = lookUpEdit1.EditValue.ToString(),
+                NgayKham = (DateTime)dtNgayKham.EditValue,
             };
 
             using (var context = new PhongKhamEntities())
@@ -147,12 +211,17 @@ namespace PhongKham.Forms_DE
         }
 
         // ################################## UPDATE #######################################
-        void CapNhatToaThuoc()
+        void capNhatToaThuoc()
         {
-            if (txtBenhDuocChanDoan.Text.Length == 0 ||
-                cbBenhNhan.SelectedIndex == -1)
+            if (txtBenhDuocChanDoan.Text.Length == 0)
             {
                 MessageBox.Show("Hãy nhập đủ thông tin.");
+                return;
+            }
+
+            if (lookUpEdit1.EditValue == "")
+            {
+                MessageBox.Show("Hãy chọn bệnh nhân.");
                 return;
             }
 
@@ -160,8 +229,9 @@ namespace PhongKham.Forms_DE
             {
                 var toaThuoc = context.tToaThuocs.FirstOrDefault(x => x.STT.ToString() == txtIdToaThuoc.Text);
                 toaThuoc.BenhDuocChanDoan = txtBenhDuocChanDoan.Text;
-                toaThuoc.BenhNhan = cbBenhNhan.SelectedValue.ToString();
-                toaThuoc.NgayKham = dtNgayKham.Value;
+                //toaThuoc.BenhNhan = cbBenhNhan.SelectedValue.ToString();
+                toaThuoc.BenhNhan = lookUpEdit1.EditValue.ToString();
+                toaThuoc.NgayKham = (DateTime)dtNgayKham.EditValue;
                 context.SaveChanges();
             }
 
@@ -170,9 +240,14 @@ namespace PhongKham.Forms_DE
             MessageBox.Show("Cập nhật toa thuốc thành công");
         }
 
-        private void groupControl1_Paint(object sender, PaintEventArgs e)
-        {
 
+
+        //######################### CHI TIET CLICK
+        private void btnEditChiTietToaThuoc_Click(object sender, EventArgs e)
+        {
+            frmChiTietToaThuoc frm = new frmChiTietToaThuoc(this);
+            frm.ShowDialog();
         }
+
     }
 }
